Huge refactor to use classes for constructs.
Time will tell if this was a good idea...
Instead of just using an enum to indicate which construct is placed on a tile,
use a class for each type of construct.

This allows managing state of the construct (e.g. ammo for turrets) and allows
multiple tiles to reference the same large construct.